requirements.txt
flask==3.0.3
flask-socketio==5.4.1
python-socketio[client]